Improve Upgrade Handling while publishing an App
Add Documentation for upgrade parameter.
If upgrade is not needed just install the app. this helps when the user just wants to have the app running but doesn't know if an upgrade is necessary or not. just always publish with the upgrade parameter and then the app is ready in any case.