Fix New-BcNuGetPackage for Runtime 3.0 (BC14) (#3307)
BC14 (runtime 3.0) does not use `id` as JSON key for the dependencies.
Instead `appId` is used. `New-BcNuGetPackage` expects `id` key, which
results in the following error:
```cmd
2024-01-24T09:17:47.2437517Z ##[error]ForEach-Object : Die Eigenschaft "id" wurde für dieses Objekt nicht gefunden. Vergewissern Sie sich, dass die 
Eigenschaft vorhanden ist.
In C:\Program Files\WindowsPowerShell\Modules\BcContainerHelper\6.0.4\NuGet\New-BcNuGetPackage.ps1:192 Zeichen:37
+             $appJson.dependencies | ForEach-Object {
+                                     ~~~~~~~~~~~~~~~~
    + CategoryInfo          : NotSpecified: (:) [ForEach-Object], PropertyNotFoundException
    + FullyQualifiedErrorId : PropertyNotFoundStrict,Microsoft.PowerShell.Commands.ForEachObjectCommand
```

Additionally the description of `packageId` is not correct anymore. In
theory you can pass anything there, but in case of dependencies existing
the template to replace placeholders is required, otherwise the ID will
be the same for all dependencies.

---------

Co-authored-by: Freddy Kristiansen <freddy.kristiansen@microsoft.com>