Wrong dependency resolution plus performance optimizations when building (#3757)
Two new parameter in Sort-AppFoldersByDependencies -
onlyTheseAppFoldersPlusDepending and SkipApps. If specified, then only
the appFolders specified in the first array plus the apps depending on
these recursively will be sorted. The remaining folder not included in
the sorted appfolders list will be returned in skipApps. This is used
for incrementals builds in AL-Go for GitHub.

Remove some excessive logging

If Import-TestToolkitToBcContainer was running with useCompilerFolder,
then it would require a container to be present at that time.

Make Run-AlPipeline do just-in-time creation of CompilerFolder and
Containers - meaning that if you are using useCompilerFolder and do not
have doNotPublishApps set - it will create a compilerFolder and compile
all apps using that, and afterwards - it will create the container,
publish the apps and run the tests.

Also...
Fixes https://github.com/microsoft/AL-Go/issues/1338

---------

Co-authored-by: freddydk <freddydk@users.noreply.github.com>