Extend condition to check for name and publisher (#3807)
We have the problem, that we have a app which is named "System" as well.
The problem is that we get the following error when we use the funciton
`Compile-AppWithBcCompilerFolder`:

![image](https://github.com/user-attachments/assets/b08c5ec4-ee42-4673-b4f4-7af67871e475)

Therefore we need to also check the publisher of the dependency to avoid
the error and wrong behaviour in the other places.