fix: ensure application family is properly used on appsource deployment (#3842)
Currently, the module is not properly forwarding `ApplicationFamily`
top/down the stack.
As a result, when we try to deploy an package from the App Source on an
environment involving in Embed Application program (where Application
Family is needed) - the script crash stating the environment is not
available.

This is caused by `$deploymentUrl` which is currently not passing the
application family - and as a result - attempt to retrieve an
environment using standard Business Central deployment.

Second issue occures while the script attempt to trigger the extension
deployment against an environment with `$installUri` which is currently
not passing the application family - and as a result - attempt to
retrieve an environment using standard Business Central deployment.

Third issue occures while the script attempt to retrieve the extension
deployment status against en environment with `$operationUri` which is
currently not passing the application family - and as a result - attempt
to retrieve an environment using standard Business Central deployment.

This Pull Request is addressing those issues.