Fixes 3831 and 3839: Local vsix file and Artifacts with symbol files without versions (#3868)
This is a reduced version of the former closed PR #3834.

- Allow vsixFile to be a local file for New-BCCompilerFolder (#3831)
- Fixes issues determining test symbols for W1 and OnPrem artifacts not
containing versions in their name (#3839)

After that, there is still an issue with RunAlTool always downloading
the AL Language extension. There should be a check or an additional
parameter as hint for an existing vsix or extracted vsix.

Regarding the other PR: I would recommend modifying the caching paths
(in AL-Go) a bit to allow caching of the single vsix in a different base
path to not mixing up caching. When specified, this is most likely a
different version compared to the rest of the compiler folder contents.