Don't remove packagesFolder if it already exists (#3910)
If compilefolders are used the packages folder will be deleted and
recreated during the pipeline run.

In certain cases especially when you use customized AL:Go with foreign
container systems, this folder exists on purpose and already includes
files. This files should not be deleted because it will lead the
pipeline to fail