Hardcode appId for System and Application (#3975)
Hardcode `appId` for _System_ and _Application_ apps.

When dependencies to _System_ and _Application_ are provided in app.json
via _system_ and _application_ properties respectively, the apps fail to
download due to an issue in the admin center API, when calling only with
app name and app publisher.
Calling the API with app ID succeeds always.

Fixes #3943