Better handling for missing or corrupt manifest files (#3985)
If a manifest ends up in a corrupt state or is deleted without the rest
of the artifact being deleted, you may run into the following error:
`Error Message: The variable '$appManifest' cannot be retrieved because
it has not been set.`

This PR adds some defensive checks to ensure that the file both exists
and is not corrupt. If it is, the artifact is redownloaded to ensure
that we have a valid manifest file.

Fixes #3981