Add support for reportSuppressedDiagnostics parameter in Run-AlPipeline (#4025)
This is related to https://github.com/microsoft/AL-Go/discussions/1911
and is a prerequisite for https://github.com/microsoft/AL-Go/pull/1978

I'm not sure if I should write the value of this parameter under the
"Parameters" section, since not all compilation settings are. What do
you think about that?

I've confirmed that it works together with
https://github.com/microsoft/AL-Go/pull/1978.

I'm also not sure if I need to add some tests for this parameter in this
repo?